backend/requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn==0.24.0

# Environment Configuration
python-decouple==3.8
python-dotenv==1.0.0

# PostgreSQL Database (simplified driver)
psycopg2-binary==2.9.9

# Cloudinary Media Management
cloudinary==1.36.0

# HTTP Client for API calls
httpx==0.25.1

# Data validation
pydantic==2.4.2
email-validator==2.1.0

# Google Sheets Integration
gspread==5.12.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0